Moving queries and responses into own helper dir
