+++ test/integration/helpers/index.js
@@ -1,0 +1,7 @@
+const queries = require('./queries');
+const responses = require('./responses');
+
+module.exports = {
+  queries,
+  responses
+};
